move commands enum closer to Cli
