+++ src/cli.rs
@@ -1,20 +1,33 @@
-use crate::commands::Commands;
+use crate::commands::{Add, List, Remove};
 use crate::config::Config;
 use crate::error::ExitCode;
-use clap::Parser;
+use clap::{Parser, Subcommand};
 use std::env;
 
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
 pub struct Cli {
     #[command(subcommand)]
-    command: Option<Commands>,
+    command: Commands,
 }
 
 #[derive(Debug)]
 pub struct ShadowedArgs {
     args: Vec<String>,
     is_raw: bool,
+}
+
+#[derive(Subcommand, Debug)]
+pub enum Commands {
+    /// Add a new alias
+    #[command(visible_alias = "a")]
+    Add(Add),
+    /// Remove an alias
+    #[command(visible_aliases = ["rm", "delete"])]
+    Remove(Remove),
+    /// List all aliases
+    #[command(visible_alias = "ls")]
+    List(List),
 }
 
 impl ShadowedArgs {
@@ -33,12 +46,10 @@ impl ShadowedArgs {
 impl Cli {
     pub fn execute(config: Config) -> ExitCode {
         let cli = Self::parse();
-        match &cli.command {
-            Some(cmd) => cmd.execute(config),
-            None => {
-                println!("Use --help for usage information");
-                ExitCode::InvalidArguments
-            }
+        match cli.command {
+            Commands::Add(cmd) => cmd.execute(config),
+            Commands::Remove(cmd) => cmd.execute(config),
+            Commands::List(cmd) => cmd.execute(config),
         }
     }
